refactor(orc8r): Remove unary package from orc8r lib
