+++ orc8r/lib/go/registry/interceptors.go
@@ -19,8 +19,15 @@ import (
 	"time"
 
 	"google.golang.org/grpc"
+	"google.golang.org/grpc/metadata"
+)
 
-	"magma/orc8r/lib/go/service/middleware/unary"
+const (
+	// Client Certificate Serial Number Header
+	CLIENT_CERT_SN_KEY = "x-magma-client-cert-serial"
+
+	// Magic Certificate Value for orc8r service clients
+	ORC8R_CLIENT_CERT_VALUE = "7ZZXAF7CAETF241KL22B8YRR7B5UF401"
 )
 
 var defaultTimeoutDuration = GrpcMaxTimeoutSec * time.Second
@@ -45,5 +52,24 @@ func TimeoutInterceptor(ctx context.Context, method string, req, resp interface{
 func CloudClientTimeoutInterceptor(ctx context.Context, method string, req, resp interface{},
 	cc *grpc.ClientConn, invoker grpc.UnaryInvoker, opts ...grpc.CallOption) error {
 
-	return TimeoutInterceptor(unary.OutgoingCloudClientCtx(ctx), method, req, resp, cc, invoker, opts...)
+	return TimeoutInterceptor(OutgoingCloudClientCtx(ctx), method, req, resp, cc, invoker, opts...)
+}
+
+// OutgoingCloudClientCtx amends outgoing cloud client context with magic client CSN metadata
+// if the original context doesn't already have client certificate serial number metadata
+func OutgoingCloudClientCtx(ctx context.Context) context.Context {
+	md, exists := metadata.FromOutgoingContext(ctx)
+	if exists {
+		if sns := md.Get(CLIENT_CERT_SN_KEY); len(sns) == 1 && len(sns[0]) > 0 {
+			// Do not alter outgoing CTX if it already has client certificate serial header
+			// this should allow dual use services (serving external as well as internal clients) to be called
+			// internally passing external client cert SN (see directoryd update servicer)
+			return ctx
+		}
+		md = md.Copy()
+		md.Set(CLIENT_CERT_SN_KEY, ORC8R_CLIENT_CERT_VALUE)
+	} else {
+		md = metadata.Pairs(CLIENT_CERT_SN_KEY, ORC8R_CLIENT_CERT_VALUE)
+	}
+	return metadata.NewOutgoingContext(ctx, md)
 }
